api_data_collector/requirements.txt
psycopg2-binary
asyncpg
websockets
pybit
okx
httpx
tenacity
pyee
python-telegram-bot
python-bitget
gate_api
cryptofeed
sqlalchemy
pandas-ta